+++ tests/dummy/app/JsonApi/Sites/Adapter.php
@@ -18,17 +18,15 @@
 namespace DummyApp\JsonApi\Sites;
 
 use CloudCreativity\LaravelJsonApi\Adapter\AbstractResourceAdapter;
-use CloudCreativity\LaravelJsonApi\Adapter\HydratesAttributesTrait;
-use CloudCreativity\LaravelJsonApi\Contracts\Object\ResourceObjectInterface;
+use CloudCreativity\LaravelJsonApi\Document\ResourceObject;
 use CloudCreativity\LaravelJsonApi\Utils\Str;
 use DummyApp\Entities\Site;
 use DummyApp\Entities\SiteRepository;
+use Illuminate\Support\Collection;
 use Neomerx\JsonApi\Contracts\Encoder\Parameters\EncodingParametersInterface;
 
 class Adapter extends AbstractResourceAdapter
 {
-
-    use HydratesAttributesTrait;
 
     /**
      * @var array
@@ -102,20 +100,31 @@ class Adapter extends AbstractResourceAdapter
     /**
      * @inheritDoc
      */
-    protected function createRecord(ResourceObjectInterface $resource)
+    protected function createRecord(ResourceObject $resource)
     {
         return new Site($resource->getId());
     }
 
     /**
+     * @inheritDoc
+     */
+    protected function fillAttributes($record, Collection $attributes)
+    {
+        foreach ($attributes as $field => $value) {
+            $this->fillAttribute($record, $field, $value);
+        }
+    }
+
+
+    /**
      * @param object $record
-     * @param string $attrKey
+     * @param string $field
      * @param mixed $value
      * @return void
      */
-    protected function hydrateAttribute($record, $attrKey, $value)
+    protected function fillAttribute($record, $field, $value)
     {
-        $method = 'set' . Str::classify($attrKey);
+        $method = 'set' . Str::classify($field);
 
         call_user_func([$record, $method], $value);
     }